[Abhishek Puri Goswami] : fix SonarCloud quality gate — resolve 5 reliability bugs and 8 security vulnerabilities

+++ auth-service/src/main/java/com/connecthub/auth/config/IpRateLimiter.java
@@ -73,7 +73,7 @@ public class IpRateLimiter {
     private boolean checkLimit(String key, int limit, int windowSeconds) {
         Long count = redis.opsForValue().increment(key);
         if (count != null && count == 1) {
-            redis.expire(key, windowSeconds + 60, TimeUnit.SECONDS);
+            redis.expire(key, (long) windowSeconds + 60, TimeUnit.SECONDS);
         }
         return count != null && count <= limit;
     }
